Compact largest tables

+++ Xtzkt.Api/Repositories/BlockRepository.cs
@@ -216,7 +216,7 @@ public class BlockRepository(
             switch (fields[i].Full)
             {
                 case "layer":
-                    foreach (var row in rows) result[j++][i] = Layers.ToString(row.Layer);
+                    foreach (var row in rows) result[j++][i] = Layers.ToString((int)row.Layer);
                     break;
                 // Block
                 case "id":
